requirements.txt
streamlit
pandas
scikit-learn
nltk
PyPDF2
python-docx
Pillow
pytesseract
plotly